+++ app/src/main/java/org/adrienbricchi/waitingformoranis/service/tmdb/TmdbError.java
@@ -16,32 +16,19 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
+package org.adrienbricchi.waitingformoranis.service.tmdb;
 
-buildscript {
+import com.fasterxml.jackson.annotation.JsonAlias;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-    repositories {
-        google()
-        mavenCentral()
-    }
 
-    dependencies {
-        classpath 'com.android.tools.build:gradle:8.0.2'
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
-        classpath 'com.google.android.gms:oss-licenses-plugin:0.10.6'
-        classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.0'
-    }
+@Getter
+@NoArgsConstructor
+public class TmdbError {
 
-}
+    private boolean success;
+    private @JsonAlias("status_code") int statusCode;
+    private @JsonAlias("status_message") String statusMessage;
 
-allprojects {
-    repositories {
-        google()
-        mavenCentral()
-    }
-}
-
-task clean(type: Delete) {
-    delete rootProject.buildDir
 }