+++ app/src/test/java/org/adrienbricchi/waitingformoranis/utils/ReleaseUtilsTest.java
@@ -1,6 +1,6 @@
 /*
  * Waiting For Moranis
- * Copyright (C) 2020-2021
+ * Copyright (C) 2020-2023
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -15,6 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
+
 package org.adrienbricchi.waitingformoranis.utils;
 
 import org.adrienbricchi.waitingformoranis.models.Movie;
@@ -40,10 +41,10 @@ public class ReleaseUtilsTest {
     public void countryLocale() {
 
         assertEquals(CANADA.getCountry(), ReleaseUtils.countryLocale("CA").getCountry());
-        Assert.assertNotNull(ReleaseUtils.countryLocale("NOTEXISTINGCOUNTRY").getCountry());
+        assertNotNull(ReleaseUtils.countryLocale("NOTEXISTINGCOUNTRY").getCountry());
 
         assertEquals("United States", US.getDisplayCountry(US));
-        assertEquals("Etats-Unis", US.getDisplayCountry(Locale.FRANCE));
+        assertEquals("États-Unis", US.getDisplayCountry(FRANCE));
     }