+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.3.5](https://github.com/adrienbricchi/waiting-for-moranis/releases/tag/1.3.5) - 2023-07-03
+### Added
+- Canceled movies parsing
+
+## [1.3.4](https://github.com/adrienbricchi/waiting-for-moranis/releases/tag/1.3.4) - 2023-04-11
+### Fixed
+- Movies release date parse from TMDB
+
+
 ## [1.3.3](https://github.com/adrienbricchi/waiting-for-moranis/releases/tag/1.3.3) - 2021-10-21
 ### Fixed
 - Movies date locale